traea 2 inicializa pgdir

+++ translate.c
@@ -7,6 +7,6 @@ int main(int argc,char *argv[])
     printf(1,"\n ----ADDRESS TRANSLATION----\n");
     int d = 123;
     printf(1,"\n Entered virtual address: %p\n",&d);
-    printf(1,"\n Status from system call: %p",addresstranslate(&d));
+    printf(1,"\n Status from system call: %p",addresstranslate((uint)(&d)));
     exit();
 }
